# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"userCreateRequest":{"$ref":"#/components/schemas/UserCreateRequest"},"profile":{"type":"string","format":"binary"}},"required":["userCreateRequest"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/readStatuses":{"get":{"tags":["read-status-controller"],"operationId":"getReadStatus","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReadStatus"}}}}}}},"post":{"tags":["read-status-controller"],"operationId":"createReadStatus","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadStatusCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReadStatus"}}}}}}},"/api/messages":{"get":{"tags":["message-controller"],"operationId":"getAllMessageChannel","parameters":[{"name":"channelId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"}}}}}}},"post":{"tags":["message-controller"],"operationId":"sendMessage","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"messageCreateRequest":{"$ref":"#/components/schemas/MessageCreateRequest"},"attachments":{"type":"array","items":{"type":"string","format":"binary"}}},"required":["messageCreateRequest"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/api/channels/public":{"post":{"tags":["channel-controller"],"operationId":"createPublicChannel","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicChannelCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChannelDto"}}}}}}},"/api/channels/private":{"post":{"tags":["channel-controller"],"operationId":"createPrivateChannel","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateChannelCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChannelDto"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/users/{userId}":{"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"patch":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"userUpdateRequest":{"$ref":"#/components/schemas/UserUpdateRequest"},"profile":{"type":"string","format":"binary"}},"required":["userUpdateRequest"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/users/{userId}/userStatus":{"patch":{"tags":["user-controller"],"operationId":"updateUserStatus","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/readStatuses/{readStatusId}":{"patch":{"tags":["read-status-controller"],"operationId":"updateReadStatus","parameters":[{"name":"readStatusId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadStatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReadStatus"}}}}}}},"/api/messages/{messageId}":{"delete":{"tags":["message-controller"],"operationId":"deleteMessage","parameters":[{"name":"messageId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}}},"patch":{"tags":["message-controller"],"operationId":"updateMessage","parameters":[{"name":"messageId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Message"}}}}}}},"/api/channels/{channelId}":{"delete":{"tags":["channel-controller"],"operationId":"deleteChannel","parameters":[{"name":"channelId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChannelDto"}}}}}},"patch":{"tags":["channel-controller"],"operationId":"updateChannel","parameters":[{"name":"channelId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicChannelUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChannelDto"}}}}}}},"/api/channels":{"get":{"tags":["channel-controller"],"operationId":"getAllChannelUser","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChannelDto"}}}}}}}},"/api/binaryContents":{"get":{"tags":["binary-content-controller"],"operationId":"findAll","parameters":[{"name":"binaryContentIds","in":"query","required":true,"schema":{"type":"array","items":{"type":"string","format":"uuid"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BinaryContent"}}}}}}}},"/api/binaryContents/{binaryContentId}":{"get":{"tags":["binary-content-controller"],"operationId":"find","parameters":[{"name":"binaryContentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BinaryContent"}}}}}}}},"components":{"schemas":{"UserCreateRequest":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"UserDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"username":{"type":"string"},"email":{"type":"string"},"profileId":{"type":"string","format":"uuid"},"online":{"type":"boolean"}}},"ReadStatusCreateRequest":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"channelId":{"type":"string","format":"uuid"},"lastReadAt":{"type":"string","format":"date-time"}}},"ReadStatus":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"userId":{"type":"string","format":"uuid"},"channelId":{"type":"string","format":"uuid"},"lastReadAt":{"type":"string","format":"date-time"}}},"MessageCreateRequest":{"type":"object","properties":{"content":{"type":"string"},"channelId":{"type":"string","format":"uuid"},"authorId":{"type":"string","format":"uuid"}}},"Message":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"content":{"type":"string"},"channelId":{"type":"string","format":"uuid"},"authorId":{"type":"string","format":"uuid"},"attachmentIds":{"type":"array","items":{"type":"string","format":"uuid"}}}},"PublicChannelCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}},"ChannelDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["PUBLIC","PRIVATE"]},"name":{"type":"string"},"description":{"type":"string"},"participantIds":{"type":"array","items":{"type":"string","format":"uuid"}},"lastMessageAt":{"type":"string","format":"date-time"}}},"PrivateChannelCreateRequest":{"type":"object","properties":{"participantIds":{"type":"array","items":{"type":"string","format":"uuid"}}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"profileId":{"type":"string","format":"uuid"}}},"UserUpdateRequest":{"type":"object","properties":{"newUsername":{"type":"string"},"newEmail":{"type":"string"},"newPassword":{"type":"string"}}},"UserStatusUpdateRequest":{"type":"object","properties":{"newLastActiveAt":{"type":"string","format":"date-time"}}},"ReadStatusUpdateRequest":{"type":"object","properties":{"newLastReadAt":{"type":"string","format":"date-time"}}},"MessageUpdateRequest":{"type":"object","properties":{"newContent":{"type":"string"}}},"PublicChannelUpdateRequest":{"type":"object","properties":{"newName":{"type":"string"},"newDescription":{"type":"string"}}},"BinaryContent":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"fileName":{"type":"string"},"size":{"type":"integer","format":"int64"},"contentType":{"type":"string"},"bytes":{"type":"string","format":"byte"}}}}}}